+++ src/plugins/ni_daq.py
@@ -1,4 +1,4 @@
-# Author: T. Onkst | Date: 08122025
+# Author: T. Onkst | Date: 03092026
 
 from __future__ import annotations